Added delete dead instances pass
